requirements.txt
#requirements file for the application .
streamlit>=1.28.0
langchain>=0.1.0
langgraph>=0.0.40
langchain-core>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.15.0
nest-asyncio>=1.5.0
asyncio>=3.4.3
openai>=1.0.0